#331 declare gameRoom message types in backend
